requirements.txt
praw
azure-storage
python-dotenv
